refactor(validate): modular the validator
