+++ src/components/config/validation/index.ts
@@ -22,10 +22,10 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
-import { validateConfigFile } from './validate-config-file'
-import { validateNotification } from './validate-notification'
-import { validateProbes } from './validate-probe'
-import { validateSymonConfig } from './validate-symon-config'
+import { validateConfigFile } from './validator/config-file'
+import { validateNotification } from './validator/notification'
+import { validateProbes } from './validator/probe'
+import { validateSymonConfig } from './validator/symon-config'
 
 export default {
   validateConfigFile,
